refactor media details paging source and delete the previous files
