+++ feature/mediaDetails/mediaDetailsUi/src/main/java/com/feature/mediaDetails/mediaDetailsUi/ui/paging/PagingSource.kt
@@ -1,17 +1,16 @@
-package com.feature.mediaDetails.mediaDetailsUi.ui.pagging
+package com.feature.mediaDetails.mediaDetailsUi.ui.paging
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
 
-abstract class BasePagingSource<Media: Any>(
-    protected val itemId: Int,
-    protected  val mediaUseCase:suspend (itemId: Int, page:Int)-> List<Media>
-): PagingSource<Int,Media>() {
+class PagingSource<Media: Any>(
+     val mediaUseCase:suspend (page:Int)-> List<Media>
+): PagingSource<Int, Media>() {
 
     override suspend fun load(params: LoadParams<Int>): LoadResult<Int, Media> {
         val page = params.key ?: 1
         return try {
-            val response = mediaUseCase(itemId,page)
+            val response = mediaUseCase(page)
             LoadResult.Page(
                 data = response as List<Media>,
                 prevKey = if (page == 1) null else page - 1,
